fix: repair the lowdb storage backend
Selecting chaite.storage=lowdb could never have worked. initChaite
imported 'storage/lowdb/storage.js' as a bare specifier, but there is no
package by that name and no imports map for it, so the branch threw
ERR_MODULE_NOT_FOUND before constructing anything. Use a relative path
and pull both exports from the single dynamic import.

Everything behind that crash was also broken:

- toolsGroupStorage was never assigned in the lowdb branch, so
  ToolsGroupManager.init() received undefined even though
  LowDBToolsGroupDTOsStorage existed and was simply never imported.
- LowDBToolsGroupDTOsStorage and LowDBUserStateStorage did not implement
  listItemsByEqFilter / listItemsByInQuery, and LowDBMcpServerStorage did
  not implement clear() — all abstract on ChaiteStorage.
- getName() was missing on five of the eight lowdb storages, so they
  logged as 'unknown'.
- OperationLogManager was only attached under sqlite, leaving lowdb users
  with no operation log at all. Attach it unconditionally; durable
  storage stays sqlite-only, since operation logs are high-frequency
  writes retained by the tens of thousands and lowdb rewrites the whole
  file per write.

Also:

- Move YunzaiUserState, a backend-agnostic DTO, out of the lowdb storage
  module and into models/chaite/user_state.js. apps/chat.js was importing
  a lowdb path regardless of the configured backend.
- Add a default branch to the storage switch so an unknown
  chaite.storage value fails with a clear message instead of a chain of
  undefined storages.
- Widen the sqlite3 peer range to >=5 <7. Miao-Yunzai pins ^6.0.1, so the
  declared range was unmet by the only supported host.
