+++ package.json
@@ -17,7 +17,7 @@
     "sqlite-vec": "^0.1.7-alpha.2"
   },
   "peerDependencies": {
-    "sqlite3": ">=5 <6"
+    "sqlite3": ">=5 <7"
   },
   "pnpm": {}
 }
